refactor(agent-platform): move FC driver primitives into shared fcvm package
Moves the Firecracker driver primitives (driver.go, launcher.go,
provisioner.go and their tests) out of fc-agentd/internal/driver into the
shared fcvm/driver package so a sibling daemon (semgrep-scand) can reuse
them: Go internal/ visibility forbids sibling imports. This completes the
fcvm extraction (after fcclient in Task 0.1), giving semgrep-scand reusable
Claim/Restore/Snapshot/provisioner primitives.

The package now imports only stdlib plus the already-shared
fcvm/fcclient and substrate packages, so there is no fc-agentd/internal
back-reference and no import cycle. Egress and control were intentionally
left in fc-agentd: the semgrep guest needs no egress.
